decoder: optimize binary reader and LEB decoding
The decoder previously paid Result binding costs for every primitive read
and maintained separate byte-array and SourceReader implementations. LEB128
decoding also crossed the reader abstraction one byte at a time, preventing
a contiguous-buffer fast path.

Unify both inputs behind BufferedWasmBinaryReader. Byte arrays decode
directly, while existing SourceReader callers use an internal refill adapter.
Primitive reads now throw only for malformed input, specialized LEB128
decoders use unrolled contiguous fast paths with a refill fallback, and
mutable scopes avoid allocating temporary context objects.

Sequential corpus decode comparisons against 1.4.7 produced:

                                  1.4.7    current   improvement
  ByteArray                       1.179 s  0.690 s   41.5%
  SourceReader compatibility path 1.247 s  0.677 s   45.7%

The public SourceReader contract remains unchanged.

+++ decoder/src/commonTest/kotlin/io/github/charlietap/chasm/decoder/decoder/type/heap/HeapTypeDecoderTest.kt
@@ -67,7 +67,7 @@ class HeapTypeDecoderTest {
             )
             val reader = FakeWasmBinaryReader(
                 fakeUByteReader = fakeUByteReader,
-                fakeS33Reader = { Ok(expectedTypeIndex) },
+                fakeS33Reader = { Ok(expectedTypeIndex.toLong()) },
                 fakePeekReader = { peekReader },
             )
             val context = decoderContext(reader)
